Lint fixes and multiple bids test.
