+++ basicswap/network.py
@@ -5,11 +5,9 @@
 # Distributed under the MIT software license, see the accompanying
 # file LICENSE or http://www.opensource.org/licenses/mit-license.php.
 
-from .interface_btc import BTCInterface
-from .chainparams import Coins
+'''
+TODO:
+'''
 
-
-class LTCInterface(BTCInterface):
-    @staticmethod
-    def coin_type():
-        return Coins.LTC
+class Peer:
+    pass
