+++ tests/lint/spelling.ignore-words.txt
@@ -1,2 +1,2 @@
 eventtypes
-
+wit
